move navigation into page files
