+++ apps/web/src/pages/AdminDashboardPage/AdminDashboardPage.tsx
@@ -1,8 +1,10 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
 import { useQuery } from "urql";
+import HomeOutlineIcon from "@heroicons/react/24/outline/HomeIcon";
+import HomeSolidIcon from "@heroicons/react/24/solid/HomeIcon";
 
-import { Heading, Pending } from "@gc-digital-talent/ui";
+import { Heading, Pending, IconType } from "@gc-digital-talent/ui";
 import { useAuthorization, hasRole } from "@gc-digital-talent/auth";
 import { User, graphql } from "@gc-digital-talent/graphql";
 
@@ -14,37 +16,43 @@ import AdminHero from "~/components/Hero/AdminHero";
 import {
   pageTitle as indexPoolPageTitle,
   pageSolidIcon as indexPoolPageIcon,
-} from "~/pages/Pools/IndexPoolPage/navigation";
+} from "~/pages/Pools/IndexPoolPage/IndexPoolPage";
 import {
   pageTitle as allPoolCandidatesPageTitle,
   pageSolidIcon as allPoolCandidatesPageIcon,
-} from "~/pages/PoolCandidates/AllPoolCandidatesPage/navigation";
+} from "~/pages/PoolCandidates/AllPoolCandidatesPage/AllPoolCandidatesPage";
 import {
   pageTitle as indexSearchRequestPageTitle,
   pageSolidIcon as indexSearchRequestPageIcon,
-} from "~/pages/SearchRequests/IndexSearchRequestPage/navigation";
+} from "~/pages/SearchRequests/IndexSearchRequestPage/IndexSearchRequestPage";
 import {
   pageTitle as indexTeamPageTitle,
   pageSolidIcon as indexTeamPageIcon,
-} from "~/pages/Teams/IndexTeamPage/navigation";
+} from "~/pages/Teams/IndexTeamPage/IndexTeamPage";
 import {
   pageTitle as indexUserPageTitle,
   pageSolidIcon as indexUserPageIcon,
-} from "~/pages/Users/IndexUserPage/navigation";
-import { indexClassificationPageSolidIcon as indexClassificationPageIcon } from "~/pages/Classifications/navigation";
-import { indexDepartmentPageSolidIcon as indexDepartmentPageIcon } from "~/pages/Departments/navigation";
+} from "~/pages/Users/IndexUserPage/IndexUserPage";
+import { pageSolidIcon as indexClassificationPageIcon } from "~/pages/Classifications/IndexClassificationPage";
+import { pageSolidIcon as indexDepartmentPageIcon } from "~/pages/Departments/IndexDepartmentPage";
 import {
-  indexSkillPageTitle,
-  indexSkillPageSolidIcon as indexSkillPageIcon,
-} from "~/pages/Skills/navigation";
+  pageTitle as indexSkillPageTitle,
+  pageSolidIcon as indexSkillPageIcon,
+} from "~/pages/Skills/IndexSkillPage";
 import {
-  indexSkillFamilyPageTitle,
-  indexSkillFamilyPageSolidIcon as indexSkillFamilyPageIcon,
-} from "~/pages/SkillFamilies/navigation";
+  pageTitle as indexSkillFamilyPageTitle,
+  pageSolidIcon as indexSkillFamilyPageIcon,
+} from "~/pages/SkillFamilies/IndexSkillFamilyPage";
 
 import LinkWell from "./components/LinkWell";
-import { pageTitle } from "./navigation";
 
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Dashboard",
+  id: "ArwIQV",
+  description: "Title for dashboard",
+});
+export const pageOutlineIcon: IconType = HomeOutlineIcon;
+export const pageSolidIcon: IconType = HomeSolidIcon;
 interface DashboardPageProps {
   currentUser?: User | null;
 }
