+++ apps/web/src/pages/Classifications/IndexClassificationPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import TagOutlineIcon from "@heroicons/react/24/outline/TagIcon";
+import TagSolidIcon from "@heroicons/react/24/solid/TagIcon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
@@ -7,7 +11,14 @@ import useRoutes from "~/hooks/useRoutes";
 import AdminHero from "~/components/Hero/AdminHero";
 
 import ClassificationTableApi from "./components/ClassificationTable";
-import { indexClassificationPageTitle as pageTitle } from "./navigation";
+
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Classifications",
+  id: "kvpRgN",
+  description: "Title for classifications",
+});
+export const pageOutlineIcon: IconType = TagOutlineIcon;
+export const pageSolidIcon: IconType = TagSolidIcon;
 
 export const IndexClassificationPage = () => {
   const intl = useIntl();
