+++ apps/web/src/pages/Departments/CreateDepartmentPage.tsx
@@ -15,9 +15,8 @@ import {
   useCreateDepartmentMutation,
 } from "~/api/generated";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
+import { pageTitle as indexDepartmentPageTitle } from "~/pages/Departments/IndexDepartmentPage";
 import AdminHero from "~/components/Hero/AdminHero";
-
-import { indexDepartmentPageTitle } from "./navigation";
 
 type FormValues = CreateDepartmentInput;
