+++ apps/web/src/pages/Departments/IndexDepartmentPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import BuildingOffice2OutlineIcon from "@heroicons/react/24/outline/BuildingOffice2Icon";
+import BuildingOffice2SolidIcon from "@heroicons/react/24/solid/BuildingOffice2Icon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
@@ -7,7 +11,14 @@ import useRoutes from "~/hooks/useRoutes";
 import AdminHero from "~/components/Hero/AdminHero";
 
 import DepartmentTableApi from "./components/DepartmentTable";
-import { indexDepartmentPageTitle as pageTitle } from "./navigation";
+
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Departments",
+  id: "+d/NdU",
+  description: "Title for departments",
+});
+export const pageOutlineIcon: IconType = BuildingOffice2OutlineIcon;
+export const pageSolidIcon: IconType = BuildingOffice2SolidIcon;
 
 export const DepartmentPage = () => {
   const intl = useIntl();
