+++ apps/web/src/pages/Departments/UpdateDepartmentPage.tsx
@@ -20,10 +20,9 @@ import {
   useUpdateDepartmentMutation,
 } from "~/api/generated";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
+import { pageTitle as indexDepartmentPageTitle } from "~/pages/Departments/IndexDepartmentPage";
 import useRequiredParams from "~/hooks/useRequiredParams";
 import AdminHero from "~/components/Hero/AdminHero";
-
-import { indexDepartmentPageTitle } from "./navigation";
 
 type FormValues = UpdateDepartmentInput;
