+++ apps/web/src/pages/PoolCandidates/AllPoolCandidatesPage/AllPoolCandidatesPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import IdentificationOutlineIcon from "@heroicons/react/24/outline/IdentificationIcon";
+import IdentificationSolidIcon from "@heroicons/react/24/solid/IdentificationIcon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import useRoutes from "~/hooks/useRoutes";
 import PoolCandidatesTable from "~/components/PoolCandidatesTable/PoolCandidatesTable";
@@ -11,7 +15,13 @@ import {
 } from "~/api/generated";
 import AdminHero from "~/components/Hero/AdminHero";
 
-import { pageTitle } from "./navigation";
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Candidate search",
+  id: "i16C7G",
+  description: "Title for the all pool candidates page",
+});
+export const pageOutlineIcon: IconType = IdentificationOutlineIcon;
+export const pageSolidIcon: IconType = IdentificationSolidIcon;
 
 export const AllPoolCandidatesPage = () => {
   const intl = useIntl();
