+++ apps/web/src/pages/SearchRequests/IndexSearchRequestPage/IndexSearchRequestPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import PaperAirplaneOutlineIcon from "@heroicons/react/24/outline/PaperAirplaneIcon";
+import PaperAirplaneSolidIcon from "@heroicons/react/24/solid/PaperAirplaneIcon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import SearchRequestTable from "~/components/SearchRequestTable/SearchRequestTable";
@@ -7,7 +11,13 @@ import useRoutes from "~/hooks/useRoutes";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
 import AdminHero from "~/components/Hero/AdminHero";
 
-import { pageTitle } from "./navigation";
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Talent requests",
+  id: "3NW70Q",
+  description: "Title for the index search request page",
+});
+export const pageOutlineIcon: IconType = PaperAirplaneOutlineIcon;
+export const pageSolidIcon: IconType = PaperAirplaneSolidIcon;
 
 export const IndexSearchRequestPage = () => {
   const intl = useIntl();
