+++ apps/web/src/pages/SkillFamilies/IndexSkillFamilyPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import CloudOutlineIcon from "@heroicons/react/24/outline/CloudIcon";
+import CloudSolidIcon from "@heroicons/react/24/solid/CloudIcon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import AdminHero from "~/components/Hero/AdminHero";
 import SEO from "~/components/SEO/SEO";
@@ -7,7 +11,14 @@ import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWr
 import useRoutes from "~/hooks/useRoutes";
 
 import SkillFamilyTableApi from "./components/SkillFamilyTable";
-import { indexSkillFamilyPageTitle as pageTitle } from "./navigation";
+
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Skill families",
+  id: "0E9hiS",
+  description: "Title for skill families",
+});
+export const pageOutlineIcon: IconType = CloudOutlineIcon;
+export const pageSolidIcon: IconType = CloudSolidIcon;
 
 const IndexSkillFamilyPage = () => {
   const intl = useIntl();
