+++ apps/web/src/pages/Skills/IndexSkillPage.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { useIntl } from "react-intl";
+import { MessageDescriptor, defineMessage, useIntl } from "react-intl";
+import BoltOutlineIcon from "@heroicons/react/24/outline/BoltIcon";
+import BoltSolidIcon from "@heroicons/react/24/solid/BoltIcon";
+
+import { IconType } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
@@ -7,7 +11,14 @@ import useRoutes from "~/hooks/useRoutes";
 
 import SkillTableApi from "./components/SkillTable";
 import AdminHero from "../../components/Hero/AdminHero";
-import { indexSkillPageTitle as pageTitle } from "./navigation";
+
+export const pageTitle: MessageDescriptor = defineMessage({
+  defaultMessage: "Skills",
+  id: "/UKT+/",
+  description: "Title for skills",
+});
+export const pageSolidIcon: IconType = BoltSolidIcon;
+export const pageOutlineIcon: IconType = BoltOutlineIcon;
 
 export const IndexSkillPage = () => {
   const intl = useIntl();
