+++ apps/web/src/pages/Users/UpdateUserPage/UpdateUserPage.tsx
@@ -35,7 +35,6 @@ import useRoutes from "~/hooks/useRoutes";
 import useRequiredParams from "~/hooks/useRequiredParams";
 import AdminContentWrapper from "~/components/AdminContentWrapper/AdminContentWrapper";
 import adminMessages from "~/messages/adminMessages";
-import { pageTitle as indexUserPageTitle } from "~/pages/Users/IndexUserPage/navigation";
 
 import UserRoleTable from "./components/IndividualRoleTable";
 import TeamRoleTable from "./components/TeamRoleTable";
